Breaking up the repository into 2 separate go modules
Namely one containing everything needed for clients to import into
repositories using this one; and then another go module for all the
rest.

See more at https://github.com/kubernetes-csi/csi-proxy/pull/7#discussion_r325323995
