+++ client/README.md
@@ -1,0 +1,5 @@
+# Client go module
+
+This separate go module is the one intended to be imported by clients that want to use the CSI-proxy.
+
+It should strive to keep as many dependencies as possible, to make it easy to import in other repositories.
