+++ client/api/errors.proto
