+++ client/go.mod
@@ -1,0 +1,10 @@
+module github.com/kubernetes-csi/csi-proxy/client
+
+go 1.12
+
+require (
+	github.com/Microsoft/go-winio v0.4.14
+	github.com/golang/protobuf v1.3.2
+	github.com/pkg/errors v0.8.1
+	google.golang.org/grpc v1.23.1
+)
