+++ client/utils.go
@@ -1,7 +1,7 @@
-package internal
+package client
 
 import (
-	"github.com/kubernetes-csi/csi-proxy/internal/apiversion"
+	"github.com/kubernetes-csi/csi-proxy/client/apiversion"
 )
 
 const (
