+++ integrationtests/apigroups/client/dummy/v1/client_generated.go
@@ -9,9 +9,9 @@ import (
 	"github.com/Microsoft/go-winio"
 	"google.golang.org/grpc"
 
+	"github.com/kubernetes-csi/csi-proxy/client"
+	"github.com/kubernetes-csi/csi-proxy/client/apiversion"
 	pb "github.com/kubernetes-csi/csi-proxy/integrationtests/apigroups/api/dummy/v1"
-	"github.com/kubernetes-csi/csi-proxy/internal"
-	"github.com/kubernetes-csi/csi-proxy/internal/apiversion"
 )
 
 const groupName = "dummy"
@@ -26,7 +26,7 @@ type wrapper struct {
 // NewClient returns a client to make calls to the dummy API group version v1.
 // It's the caller's responsibility to Close the client when done.
 func NewClient() (*wrapper, error) {
-	pipePath := internal.PipePath(groupName, version)
+	pipePath := client.PipePath(groupName, version)
 
 	connection, err := grpc.Dial(pipePath,
 		grpc.WithContextDialer(func(context context.Context, s string) (net.Conn, error) {
