+++ internal/server/filesystem/api_group_generated.go
@@ -3,7 +3,7 @@
 package dummy
 
 import (
-	"github.com/kubernetes-csi/csi-proxy/internal/apiversion"
+	"github.com/kubernetes-csi/csi-proxy/client/apiversion"
 	"github.com/kubernetes-csi/csi-proxy/internal/server"
 	"github.com/kubernetes-csi/csi-proxy/internal/server/filesystem/internal"
 	"github.com/kubernetes-csi/csi-proxy/internal/server/filesystem/internal/v1alpha1"
