+++ internal/server/filesystem/internal/v1alpha1/conversion_generated.go
@@ -3,8 +3,8 @@
 package v1alpha1
 
 import (
-	"github.com/kubernetes-csi/csi-proxy/api"
-	pb "github.com/kubernetes-csi/csi-proxy/api/filesystem/v1alpha1"
+	"github.com/kubernetes-csi/csi-proxy/client/api"
+	pb "github.com/kubernetes-csi/csi-proxy/client/api/filesystem/v1alpha1"
 	"github.com/kubernetes-csi/csi-proxy/internal/server/filesystem/internal"
 )
