+++ internal/server/types.go
@@ -3,7 +3,7 @@ package server
 import (
 	"google.golang.org/grpc"
 
-	"github.com/kubernetes-csi/csi-proxy/internal/apiversion"
+	"github.com/kubernetes-csi/csi-proxy/client/apiversion"
 )
 
 type VersionedAPI struct {
